Immutable fields when being marked as touch
**Problem**

Merge applies changes to the first object, which was originally jus the
field.

This meant that React wasn't re-rendering components that referenced
these fields directly, as although properties of the object had changed,
it was still the same object.

**Solution**

Pass in an empty object as the first param to merge, so that a new
object is returned, React will re-render, and everybody is happy.